Code markups and minor refactorings
